requirements.txt
Flask==3.0.0
python-docx==1.1.0
openpyxl==3.1.2
lxml==4.9.3
gunicorn==21.2.0
